Remove casts

+++ src/lnb_exec.c
@@ -306,9 +306,9 @@ static char * __lnb_get_target_link_path (
 				dirname_len = 0;
 			}
 # ifdef HAVE_MALLOC
-			__lnb_newlinkpath = (char *) malloc ((size_t)(
+			__lnb_newlinkpath = (char *) malloc (
 				dirname_len + 1
-				+ (size_t)lsize + 1));
+				+ (size_t)lsize + 1);
 			if ( __lnb_newlinkpath == NULL )
 			{
 				break;
@@ -332,9 +332,9 @@ static char * __lnb_get_target_link_path (
 				different directory (there was a slash in the original path)
 				- append the link's directory name */
 # ifdef HAVE_MALLOC
-				__lnb_newlinkdir = (char *) malloc ((size_t)(
+				__lnb_newlinkdir = (char *) malloc (
 					dirname_len + 1
-					+ (size_t)lsize + 1));
+					+ (size_t)lsize + 1);
 				if ( __lnb_newlinkdir == NULL )
 				{
 					free (__lnb_newlinkpath);
